Phase 4.3: extract prompt DB pure helpers
